eliminated traces of old ReaderTagToolbar
