+++ WordPress/src/androidTest/java/org/wordpress/android/mocks/OAuthAuthenticatorEmptyMock.java
@@ -1,8 +1,8 @@
 package org.wordpress.android.mocks;
 
+import org.wordpress.android.models.AccountHelper;
 import org.wordpress.android.networking.AuthenticatorRequest;
 import org.wordpress.android.networking.OAuthAuthenticator;
-import org.wordpress.android.models.AccountHelper;
 
 public class OAuthAuthenticatorEmptyMock extends OAuthAuthenticator {
     public void authenticate(AuthenticatorRequest request) {
