+++ WordPress/src/androidTest/java/org/wordpress/android/mocks/XMLRPCClientCustomizableXMLMock.java
@@ -12,7 +12,6 @@ import org.xmlrpc.android.XMLRPCFault;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
 
 public class XMLRPCClientCustomizableXMLMock extends XMLRPCClientCustomizableMockAbstract {
