+++ WordPress/src/androidTest/java/org/wordpress/android/ui/notifications/NotesParseTest.java
@@ -1,6 +1,5 @@
 package org.wordpress.android.ui.notifications;
 
-import android.text.SpannableStringBuilder;
 import android.text.Spanned;
 
 import junit.framework.TestCase;
