+++ WordPress/src/main/AndroidManifest.xml
@@ -244,10 +244,6 @@
             android:theme="@style/Calypso.NoActionBar"
             android:label="@string/reader" />
         <activity
-            android:name=".ui.reader.ReaderPostListActivity1"
-            android:theme="@style/Calypso.NoActionBar"
-            android:label="@string/reader" />
-        <activity
             android:name=".ui.reader.ReaderPostPagerActivity"
             android:theme="@style/Calypso.NoActionBar"
             android:label="@string/reader_title_post_detail"
