--- WordPress/src/main/res/layout/reader_tag_toolbar.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    "toolbar" at the top of the reader post list
--->
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginBottom="@dimen/margin_small"
-    android:layout_marginLeft="@dimen/reader_card_margin"
-    android:layout_marginRight="@dimen/reader_card_margin"
-    android:layout_marginTop="@dimen/margin_small"
-    app:popupTheme="@style/ThemeOverlay.AppCompat.Light">
-
-    <org.wordpress.android.widgets.WPTextView
-        android:id="@+id/text_tag"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:background="?android:selectableItemBackground"
-        android:drawablePadding="@dimen/margin_medium"
-        android:drawableRight="@drawable/reader_dropdown_arrow"
-        android:ellipsize="end"
-        android:paddingBottom="@dimen/margin_medium"
-        android:paddingLeft="@dimen/reader_card_content_padding"
-        android:paddingRight="@dimen/reader_card_content_padding"
-        android:paddingTop="@dimen/margin_medium"
-        android:textColor="@color/grey"
-        android:textSize="@dimen/text_sz_large"
-        tools:text="current tag" />
-
-    <ImageView
-        android:id="@+id/btn_edit_tags"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentRight="true"
-        android:layout_centerVertical="true"
-        android:background="?android:selectableItemBackground"
-        android:padding="@dimen/margin_large"
-        android:contentDescription="@string/reader_menu_tags"
-        android:src="@drawable/my_site_icon_settings" />
-
-</RelativeLayout>
