--- WordPress/src/main/res/layout/reader_tag_toolbar_menu_item.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<org.wordpress.android.widgets.WPTextView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/text"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:ellipsize="end"
-    android:gravity="center_vertical"
-    android:padding="@dimen/margin_large"
-    android:textColor="@color/grey"
-    android:textSize="@dimen/text_sz_large"
-    tools:text="tag name" />
